@@ -25,11 +25,11 @@
     "build": "lerna run --scope @firebase/* --scope firebase --scope rxfire build",
     "build:exp": "lerna run --scope @firebase/*-exp --scope firebase-exp build",
     "build:release": "lerna run --scope @firebase/* --scope firebase --ignore @firebase/*-exp --ignore firebase-exp prepare",
-    "build:exp:release": "lerna run --scope @firebase/*-exp --scope firebase-exp prepare && yarn --cwd packages-exp/app-exp typings:public",
+    "build:exp:release": "yarn --cwd packages/app build:deps && lerna run --scope @firebase/*-exp --scope firebase-exp prepare && yarn --cwd packages-exp/app-exp typings:public",
     "link:packages": "lerna exec --scope @firebase/* --scope firebase --scope rxfire -- yarn link",
     "stage:packages": "./scripts/prepublish.sh",
     "repl": "node tools/repl.js",
-    "release": "node scripts/release/cli.js",
+    "release": "ts-node-script scripts/release/cli.ts",
     "pretest": "node tools/pretest.js",
     "test": "lerna run --concurrency 4 --stream test",
     "test:ci": "lerna run --concurrency 4 --stream test:ci",
@@ -60,10 +60,15 @@
     "integration/*"
   ],
   "devDependencies": {
+    "@changesets/changelog-github": "0.2.5",
+    "@changesets/cli": "2.9.1",
     "@microsoft/api-documenter": "7.7.20",
     "@microsoft/api-extractor": "7.7.13",
     "@types/chai": "4.2.11",
     "@types/chai-as-promised": "7.1.2",
+    "@types/child-process-promise": "2.2.1",
+    "@types/clone": "0.1.30",
+    "@types/inquirer": "6.5.0",
     "@types/long": "4.0.1",
     "@types/mocha": "7.0.2",
     "@types/node": "12.12.37",
@@ -71,24 +76,25 @@
     "@types/sinon-chai": "3.2.4",
     "@types/tmp": "0.2.0",
     "@types/yargs": "15.0.4",
+    "@types/listr": "0.14.2",
     "@typescript-eslint/eslint-plugin": "2.30.0",
     "@typescript-eslint/eslint-plugin-tslint": "2.30.0",
     "@typescript-eslint/parser": "2.30.0",
     "babel-loader": "8.1.0",
     "chai": "4.2.0",
     "chai-as-promised": "7.1.1",
-    "chalk": "3.0.0",
+    "chalk": "4.1.0",
     "child-process-promise": "2.2.1",
     "clone": "2.1.2",
     "coveralls": "3.1.0",
     "del": "5.1.0",
     "dependency-graph": "0.9.0",
-    "eslint": "6.8.0",
+    "eslint": "7.3.1",
     "eslint-plugin-import": "2.20.2",
     "express": "4.17.1",
     "find-free-port": "2.0.0",
     "firebase-functions": "3.6.1",
-    "firebase-tools": "7.16.2",
+    "firebase-tools": "8.4.3",
     "git-rev-sync": "2.0.0",
     "glob": "7.1.6",
     "http-server": "0.12.3",
