@@ -16,30 +16,33 @@
  */
 
 import { expect } from 'chai';
+import { resolve } from 'path';
 
 import { extractDependencies } from '../../../../scripts/exp/extract-deps.helpers';
 
 import * as dependencies from './dependencies.json';
-import * as pkg from '../../package.json';
+import * as pkg from '../package.json';
 import { forEach } from '../../src/util/obj';
 
 describe('Dependencies', () => {
   forEach(dependencies, (api, { dependencies }) => {
     it(api, () => {
-      return extractDependencies(api, pkg.exp).then(extractedDependencies => {
-        expect(extractedDependencies.classes).to.have.members(
-          dependencies.classes,
-          'for classes'
-        );
-        expect(extractedDependencies.functions).to.have.members(
-          dependencies.functions,
-          'for functions'
-        );
-        expect(extractedDependencies.variables).to.have.members(
-          dependencies.variables,
-          'for variables'
-        );
-      });
+      return extractDependencies(api, resolve('lite', pkg.main)).then(
+        extractedDependencies => {
+          expect(extractedDependencies.classes).to.have.members(
+            dependencies.classes,
+            'for classes'
+          );
+          expect(extractedDependencies.functions).to.have.members(
+            dependencies.functions,
+            'for functions'
+          );
+          expect(extractedDependencies.variables).to.have.members(
+            dependencies.variables,
+            'for variables'
+          );
+        }
+      );
     });
   });
 });
