@@ -21,13 +21,13 @@ import { _FirebaseNamespace } from '@firebase/app-types/private';
 import { Component, ComponentType, Provider } from '@firebase/component';
 import {
   CACHE_SIZE_UNLIMITED,
-  Firestore,
+  CollectionReference,
   DocumentReference,
   DocumentSnapshot,
-  QueryDocumentSnapshot,
+  Firestore,
   Query,
+  QueryDocumentSnapshot,
   QuerySnapshot,
-  CollectionReference,
   Transaction,
   WriteBatch
 } from './api/database';
@@ -35,61 +35,23 @@ import { Blob } from './api/blob';
 import { FieldPath } from './api/field_path';
 import { GeoPoint } from './api/geo_point';
 import { Timestamp } from './api/timestamp';
-import { makeConstructorPrivate } from './util/api';
 import { FieldValue } from './api/field_value';
 
-// Public instance that disallows construction at runtime. Note that this still
-// allows instanceof checks.
-export const PublicFirestore = makeConstructorPrivate(
-  Firestore,
-  'Use firebase.firestore() instead.'
-);
-export const PublicTransaction = makeConstructorPrivate(
-  Transaction,
-  'Use firebase.firestore().runTransaction() instead.'
-);
-export const PublicWriteBatch = makeConstructorPrivate(
-  WriteBatch,
-  'Use firebase.firestore().batch() instead.'
-);
-export const PublicDocumentReference = makeConstructorPrivate(
-  DocumentReference,
-  'Use firebase.firestore().doc() instead.'
-);
-export const PublicDocumentSnapshot = makeConstructorPrivate(DocumentSnapshot);
-export const PublicQueryDocumentSnapshot = makeConstructorPrivate(
-  QueryDocumentSnapshot
-);
-export const PublicQuery = makeConstructorPrivate(Query);
-export const PublicQuerySnapshot = makeConstructorPrivate(QuerySnapshot);
-export const PublicCollectionReference = makeConstructorPrivate(
-  CollectionReference,
-  'Use firebase.firestore().collection() instead.'
-);
-export const PublicFieldValue = makeConstructorPrivate(
-  FieldValue,
-  'Use FieldValue.<field>() instead.'
-);
-export const PublicBlob = makeConstructorPrivate(
-  Blob,
-  'Use Blob.fromUint8Array() or Blob.fromBase64String() instead.'
-);
-
 const firestoreNamespace = {
-  Firestore: PublicFirestore,
+  Firestore,
   GeoPoint,
   Timestamp,
-  Blob: PublicBlob,
-  Transaction: PublicTransaction,
-  WriteBatch: PublicWriteBatch,
-  DocumentReference: PublicDocumentReference,
-  DocumentSnapshot: PublicDocumentSnapshot,
-  Query: PublicQuery,
-  QueryDocumentSnapshot: PublicQueryDocumentSnapshot,
-  QuerySnapshot: PublicQuerySnapshot,
-  CollectionReference: PublicCollectionReference,
+  Blob,
+  Transaction,
+  WriteBatch,
+  DocumentReference,
+  DocumentSnapshot,
+  Query,
+  QueryDocumentSnapshot,
+  QuerySnapshot,
+  CollectionReference,
   FieldPath,
-  FieldValue: PublicFieldValue,
+  FieldValue,
   setLogLevel: Firestore.setLogLevel,
   CACHE_SIZE_UNLIMITED
 };
