@@ -19,7 +19,7 @@ import { expect } from 'chai';
 import {
   changesFromSnapshot,
   DocumentSnapshot,
-  Firestore
+  QueryDocumentSnapshot
 } from '../../../src/api/database';
 import { Query } from '../../../src/core/query';
 import { View } from '../../../src/core/view';
@@ -34,6 +34,7 @@ import {
   orderBy,
   path
 } from '../../util/helpers';
+import { firestore } from '../../util/api_helpers';
 
 describe('DocumentChange:', () => {
   function expectPositions(
@@ -54,10 +55,17 @@ describe('DocumentChange:', () => {
     const actual = documentSetAsArray(initialSnapshot.docs);
 
     const changes = changesFromSnapshot(
-      {} as Firestore,
-      true,
       updatedSnapshot,
-      /* converter= */ null
+      true,
+      (doc, fromCache, hasPendingWrite) =>
+        new QueryDocumentSnapshot(
+          firestore(),
+          doc.key,
+          doc,
+          fromCache,
+          hasPendingWrite,
+          /* converter= */ null
+        )
     );
 
     for (const change of changes) {
