@@ -24,10 +24,6 @@ import * as request from 'request';
 // @ts-ignore
 import * as tmp from 'tmp';
 
-export interface ChildProcessPromise extends Promise<void> {
-  childProcess: ChildProcess;
-}
-
 export abstract class Emulator {
   binaryPath: string | null = null;
   emulator: ChildProcess | null = null;
@@ -72,9 +68,14 @@ export abstract class Emulator {
       if (!this.binaryPath) {
         throw new Error('You must call download() before setUp()');
       }
-      const promise: ChildProcessPromise = spawn(
+      const promise = spawn(
         'java',
-        ['-jar', path.basename(this.binaryPath), '--port', this.port],
+        [
+          '-jar',
+          path.basename(this.binaryPath),
+          '--port',
+          this.port.toString()
+        ],
         {
           cwd: path.dirname(this.binaryPath),
           stdio: 'inherit'
